+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 * nostr: add `NostrWalletConnectURI::parse` ([Yuki Kishimoto])
 * nostr: add `JobFeedbackData` struct ([Yuki Kishimoto])
 * nostr: add `EventBuilder::pow` method ([Yuki Kishimoto])
+* database: add `DatabaseHelper::fast_query` ([Yuki Kishimoto])
 * bindings(nostr): expose `as_pretty_json` for some structs ([Yuki Kishimoto])
 * bindings(sdk): expose `Client::fetch_metadata` ([Yuki Kishimoto])
 * ffi(nostr): expose `Kind::is_*` methods ([Yuki Kishimoto])